refactor view model and repository
